requirements.txt
scipy
scikit-learn
mysql-connector-python
joblib
python-dotenv
pandas
numpy
